+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.8.0 (unreleased)
 
 - Added support for inline filtering with HNSW
+- Added support for iterative index scans
 - Added casts for arrays to `sparsevec`
 - Improved cost estimation
 - Improved performance of HNSW inserts and on-disk index builds